requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
djangorestframework-simplejwt>=5.3,<6.0
mysqlclient>=2.2,<3.0
